+++ frontend/src/modules/uploader/views/uploaderItemVw.ts
@@ -1,7 +1,15 @@
+/**
+ * Saito - The Threaded Web Forum
+ *
+ * @copyright Copyright (c) the Saito Project Developers
+ * @link https://github.com/Schlaefer/Saito
+ * @license http://opensource.org/licenses/MIT
+ */
+
 import { Model } from 'backbone';
 import { View } from 'backbone.marionette';
 import App from 'models/app';
-import * as _ from 'underscore';
+import { defaults } from 'underscore';
 import AudioTpl from '../templates/uploadItemAudioTpl.html';
 import GenericTpl from '../templates/uploadItemGenericTpl.html';
 import ImageTpl from '../templates/uploadItemImageTpl.html';
@@ -11,14 +19,14 @@ import UploaderItemFooterVw from './uploaderItemFooterVw';
 
 class UploaderItemVw extends View<Model> {
     public constructor(options: IUploaderOptions) {
-        _.defaults(options, {
+        options = defaults(options, {
             className: 'card imageUploader-card',
             regions: {
                 footer: '.js-footer',
                 rgForm: '.js-rgForm',
             },
         });
-        super(...arguments);
+        super(options);
     }
 
     public getTemplate() {
@@ -48,10 +56,6 @@ class UploaderItemVw extends View<Model> {
             actionView.model = this.model;
             this.showChildView('rgForm', actionView);
         }
-
-        //// delay display of loading spinner
-        this.$('.image-uploader-spinner').css('visibility', 'hidden');
-        _.delay(() => { this.$('.image-uploader-spinner').css('visibility', 'visible'); }, 2000);
     }
 }